We need to use clones of the original objects we use for computing history.
This was causing really weird bugs as we modified/deleted critical properties of some objects. This wasn't needed before because before we got a copy "for free" when sending this over localhost to a separate NodeJS app.

Minor: HistoryManager now also stores the newly processed instance in its own memory.